Correct the loading of data, changing struct char became int

+++ SRC/graph.c
@@ -51,7 +51,7 @@ Graph new_graph(const int number_of_node) {
 void free_graph(Graph g) {
     int i = 0;
 
-	for( i = 0; i < (int) g->count_nodes; i++) {
+	for( i = 0; i < g->count_nodes; i++) {
          free(g->nodes[i]->cost);
 	     free(g->nodes[i]->subnodes);
 	     free(g->nodes[i]);
@@ -68,11 +68,11 @@ void print_graph(const Graph g) {
     
     printf("\nPrint the graph : \n");
     
-    for( i = 0; i < (int) g->count_nodes; i++) {
-        printf("***** Node name : %d \n", (int) g->nodes[i]->name);
+    for( i = 0; i < g->count_nodes; i++) {
+        printf("***** Node name : %d \n", g->nodes[i]->name);
           
-        for( j = 0; j < (int) g->nodes[i]->count_subnodes; j++) {
-            printf("\tNode Name : %d [ %d ] (%d)\n", (int) g->nodes[i]->subnodes[j]->name, (int)g->nodes[i]->cost[j], (int)g->nodes[i]->colored);
+        for( j = 0; j < g->nodes[i]->count_subnodes; j++) {
+            printf("\tNode Name : %d [ %d ] (%d)\n", g->nodes[i]->subnodes[j]->name, g->nodes[i]->cost[j], g->nodes[i]->colored);
         }
     }
 }
